Bump contact version, pass file name to server when creating Connection.
- Using the latest version of contact allows query string parameters to
be supplied to the Dial function. This lets us pass data such as the
file name (and in the future, other data that might be useful) to the
server before creating the websocket connection. This is useful since a
POST request cannot be converted to a websocket, and passing the data
through the connection after it was created would require special logic
to handle receiving the data over the connection immediately after it was
established.